requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
python-dotenv==1.0.1